Add files via upload
Changed interpretation of PCA clustering graphs as per GradeScope comments. Was not able to push through R Studio like I was able to in Lab 13. In Lab 13 I could not figure out how to upload the images, and I am still having trouble with that. Thus, I chose to just upload the PDF of the project.