Add files via upload
Same as Project 1 opted to upload as pdf in order to see graphs and visualizations. Edited the project after reviewing comments on GradeScope, edited things such as the interpretation of EDA, changed the bonferroni and type 1 error due to wrong number of tests being used, changed the linear regression model, as well as changing the logistic regression.